+++ nexus/lang/core/.gitignore
