+++ nexus/lang/core/embed-nexus.bin
