+++ nexus/lang/examples/train.c
